data/股票程式開發/stockbuddy-backend/requirements.txt
# FastAPI 框架
fastapi>=0.109.0
uvicorn>=0.27.0

# 資料處理 (使用較新版本支援 Python 3.11+)
pandas>=2.2.0
numpy>=1.26.0

# HTTP 請求
httpx>=0.26.0
aiohttp>=3.9.0

# 股票資料
yfinance>=0.2.36

# 快取
cachetools>=5.3.0

# 環境變數
python-dotenv>=1.0.0

# CORS
python-multipart>=0.0.6

# Excel 匯出 (V10.15)
openpyxl>=3.1.0

# 時區處理 (V10.24 美股支援)
pytz>=2024.1

# V10.36: ML 預測引擎 (可選，若不需要 ML 功能可不安裝)
xgboost>=2.0.0
scikit-learn>=1.4.0

# V10.37: 安全性增強
slowapi>=0.1.9

# V10.38: 資料庫支援
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# V10.38: JWT 身份驗證
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# V10.38: 錯誤監控
sentry-sdk[fastapi]>=1.40.0
